Use BSD syntax (ps aux).
"ps -aux" is not a valid command but it may be (typically is)
interpreted as the famous BSD "ps aux" syntax.

"-aux" is invalid because -x is not a GNU flag and -u requires an
argument. If x where treated as the argument to -u the command's
semantics would be changed.