Do not send ps' column headers.
ps.php naively sends ps' complete output, including the column headers.
These aren't data and are already defined in the table anyway.

A more robust solution would probably be to map the headers to the
table since they conveniently identify the data. But this was easier
and we don't need to worry too much about robustness yet.